+++ original/forms.py
@@ -19,7 +19,7 @@ from wtforms.fields import (
 def validate_comment_code(form):
     """Check provided code corresponds to checksum."""
 
-    code_checksum = filter_antispam_code(form.commentspamcheck.data)
+    code_checksum = compute_antispam_checksum(form.commentspamcheck.data)
 
     if code_checksum != form.commentspamchecksum.data:
         form.commentspamcheck.errors.append(_('Invalid validation code'))
@@ -28,12 +28,16 @@ def validate_comment_code(form):
         return True
 
 
-def filter_antispam_code(code):
+def compute_antispam_checksum(code):
     """Compute checksum of `code`."""
-    if code.startswith('$5$'):
-        return code
-    else:
-        return '$5$' + hashlib.md5(code.encode('utf-8')).hexdigest()
+    return hashlib.md5(code.encode('utf-8')).hexdigest()
+
+
+class AntispamField(HiddenField):
+    """Antispam checksum generator field."""
+
+    def _value(self):
+        return compute_antispam_checksum(self.object_data)
 
 
 class CommentForm(FlaskForm):
@@ -42,9 +46,8 @@ class CommentForm(FlaskForm):
     username = StringField(_('Name:'), [validators.DataRequired()])
     remember = BooleanField(_('Remember Name:'))
 
-    commentspamchecksum = HiddenField(
+    commentspamchecksum = AntispamField(
         default=lambda : u'{:04d}'.format(random.randint(0, 9999)),
-        filters=[filter_antispam_code],
     )
     commentspamcheck = StringField(_('Retype PIN Above:'))