+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 * Switch to flask-classful.
 * Fix antispam code generation, closes #3.
+* Move antispam code logic to a custom Field.
 
 1.3.1 (2017-03-13)
 ------------------